chore: isolate DRA utilities
